Minor door patch
Entering the door on the last level no longer increases global.currentLevel to be one more than the total number of levels. (Probably fixing an irrelevant issue but might be relevant later)

Note: Last patch added functionality to the F1 key so that it also resets save data, along with restarting the game.